pyproject.toml
[project]
name = "auto-download-estrategia-concurso"
version = "1.0.0"
description = "Baixador automático de cursos do Estratégia Concursos"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "seleniumbase>=4.0.0",
    "requests>=2.28.0",
    "tqdm>=4.65.0",
    "colorama>=0.4.6",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.2.0",
]

[project.scripts]
download-estrategia = "main:main"
